Complete Lab2
