+++ Lab2/kernel/main.c
@@ -1,6 +1,9 @@
 #include "uart.h"
 #include "shell.h"
+#include "dtb.h"
+
 #define MEMORY_POOL_SIZE 1024
+
 
 static char memory_pool[MEMORY_POOL_SIZE];
 static char *next_free = memory_pool;
@@ -15,11 +18,12 @@ void *simple_alloc(int size) {
 }
 
 
-void main()
+void main(void *dtb)
 {
     // set up serial console
     uart_init();
-    
+
+    fdt_tranverse(dtb, "linux,initrd-start", initramfs_callback);
     // say hello
     uart_puts("Booted, start testing simple_alloc\n");
