+++ Lab2/kernel/uart.h
@@ -3,3 +3,4 @@ void uart_send(unsigned int c);
 char uart_getc();
 void uart_puts(char *s);
 void uart_hex(unsigned int d);
+void uart_puts2(char *s);
