--- Lab2/kernel_dtb/config.txt
@@ -1,2 +1,0 @@
-arm_64bit=1
-initramfs initramfs.cpio 0x20000000
