--- Lab2/kernel_dtb/uart.h
@@ -1,6 +1,0 @@
-void uart_init();
-void uart_send(unsigned int c);
-char uart_getc();
-void uart_puts(char *s);
-void uart_hex(unsigned int d);
-void uart_puts2(char *s);
